refactor(www): carve out playground/ slice
Move playground UI components and entries out of app/playground/
into a top-level playground/ slice. Route files in app/playground/
become thin shells that import from @/playground/*.

- playground/components/ ← app/playground/components/ + component-loader
- playground/entries/ ← app/playground/registry/ (flattened)
- Add playground/**/* to tailwind content glob
