+++ apps/www/app/playground/layout.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { SidebarProvider } from '@/registry/ui/sidebar';
-import { PlaygroundSidebar } from './components/playground-sidebar';
+import { PlaygroundSidebar } from '@/playground/components/playground-sidebar';
 import { Toaster } from '@/registry/ui/sonner';
 
 export default function PlaygroundLayout({ children }: { children: React.ReactNode }) {
