+++ apps/www/playground/components/global-controls.tsx
