+++ apps/www/playground/entries/index.ts
@@ -27,67 +27,67 @@ export const registry: Record<string, PlaygroundEntry[]> = {
     {
       name: 'Accordion',
       slug: 'accordion',
-      load: () => import('./entries/accordion'),
+      load: () => import('./accordion'),
     },
     {
       name: 'Alert Dialog',
       slug: 'alert-dialog',
-      load: () => import('./entries/alert-dialog'),
+      load: () => import('./alert-dialog'),
     },
     {
       name: 'Avatar',
       slug: 'avatar',
-      load: () => import('./entries/avatar'),
+      load: () => import('./avatar'),
     },
     {
       name: 'Avatar Group',
       slug: 'avatar-group',
-      load: () => import('./entries/avatar-group'),
+      load: () => import('./avatar-group'),
     },
     {
       name: 'Button',
       slug: 'button',
-      load: () => import('./entries/button'),
+      load: () => import('./button'),
     },
     {
       name: 'Badge',
       slug: 'badge',
-      load: () => import('./entries/badge'),
+      load: () => import('./badge'),
     },
     {
       name: 'Button Group',
       slug: 'button-group',
-      load: () => import('./entries/button-group'),
+      load: () => import('./button-group'),
     },
     {
       name: 'Checkbox',
       slug: 'checkbox',
-      load: () => import('./entries/checkbox'),
+      load: () => import('./checkbox'),
     },
     {
       name: 'Checkbox Group',
       slug: 'checkbox-group',
-      load: () => import('./entries/checkbox-group'),
+      load: () => import('./checkbox-group'),
     },
     {
       name: 'Dropdown Menu',
       slug: 'dropdown-menu',
-      load: () => import('./entries/dropdown-menu'),
+      load: () => import('./dropdown-menu'),
     },
     {
       name: 'Input',
       slug: 'input',
-      load: () => import('./entries/input'),
+      load: () => import('./input'),
     },
     {
       name: 'Input Group',
       slug: 'input-group',
-      load: () => import('./entries/input-group'),
+      load: () => import('./input-group'),
     },
     {
       name: 'Tabs',
       slug: 'tabs',
-      load: () => import('./entries/tabs'),
+      load: () => import('./tabs'),
     },
   ],
   blocks: [],
