+++ apps/www/playground/entries/input-group.tsx
@@ -10,9 +10,9 @@ import {
   InputGroupTextarea,
 } from '@/registry/ui/input-group';
 import { ToolbarSeparator } from '@/registry/ui/toolbar';
-import { PreviewShell } from '../../components/preview-shell';
-import { SelectControl, CheckboxControl, TextControl } from '../../components/controls';
-import { INPUT_VARIANTS } from '../../components/constants';
+import { PreviewShell } from '@/playground/components/preview-shell';
+import { SelectControl, CheckboxControl, TextControl } from '@/playground/components/controls';
+import { INPUT_VARIANTS } from '@/playground/components/constants';
 import { MagnifyingGlassIcon, EnvelopeIcon, CalendarIcon, LinkIcon } from '@phosphor-icons/react';
 
 const ALIGN_OPTIONS = [
@@ -271,8 +271,9 @@ function InputGroupVariants() {
   );
 }
 
-export default {
+const entry = {
   Preview: InputGroupPreview,
   Variants: InputGroupVariants,
   defaults,
 };
+export default entry;
